# spring_rest/basic_rest/productappv8_swagger2/src/main/resources/api.yaml
openapi: 3.0.3
info:
  title: Product Service API Busycoder
  description: This is the REST API for Product Service
  license:
    name: Apache 2.0
  version: v0.0.1
externalDocs:
  description: You can refer to the Product Service Wiki Documentation
  url: https://product-service-dummy-url.com/docs
servers:
  - url: http://localhost:9090
    description: Generated server url
paths:
  /api/products/{id}:
    get:
      tags:
        - product-controller
      operationId: getById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    put:
      tags:
        - product-controller
      operationId: updateProduct
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Product"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      tags:
        - product-controller
      operationId: deleteById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: string
  /api/products:
    get:
      tags:
        - product-controller
      operationId: getAll
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      tags:
        - product-controller
      operationId: addProduct
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Product"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /hello2:
    get:
      tags:
        - hello-controller
      operationId: hello2
      parameters:
        - name: name
          in: query
          required: true
          schema:
            type: string
        - name: city
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: string
  /hello/{name}/{city}:
    get:
      tags:
        - hello-controller
      operationId: hello
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
        - name: city
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: string
components:
  schemas:
    Product:
      type: object
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
        price:
          type: number
          format: double
      required:
        - name
        - price